TXM-7846 Sometimes the on screen keyboard opens then immediately closes (#387)
This change insures that the tabtip process (on screen keyboard) is no longer running when attempting to open the keyboard.  Made another attempt to set the startup value for the TabletInputSefvice from Manual to Automatic, in the registry, could read and write without exception, but the value was unable to be set.