TXM-8652 - limits page currency fields have wrong currency if we go back and change (#566)
Turns out the fields do update, depending on whether the number in the fields were changed when we go back and forth. So the currency and the fields were working properly, it was just a redraw problem. If we force the currency input fields to redraw when they are rendered, then this goes away.