TXM-8195 *Ticket Printing* Displayed for a while instead of Handpay l… (#564)
…ockup

Fixed issues triggered by an error while printing that would cause
problems with the print transitioning to completion and the resulting
handpay. Issue is most easily reproduced by opening the printer chassis
while a cash-out voucher is printing.....this usually results in the
ticket continuing to print after the printer has reported the chassis
open. We only need to cancel the print and transition to handpay if we
do not get status notification from the print in progress that the
region of interest (human-readable barcode) has printed.