TXM-8662 Epic TTL printer stuck in handpay (#574)
For Epic TTL printers, this corrects the evaluation of the print status
so that a print properly transitions to completion. Prior to this fix,
if you caused a print error (chassis open, head open, jam) while
printing a voucher and the system triggered a handpay (region of
interest not printed), once the print error was resolved and the handpay
keyed-off, a handpay receipt would print but since the printing of the
handpay receipt was also not properly transitioning to complete, the
handpay would trigger again.....which would continue over and over, so
stuck in handpay.

Another subtle visual side effect of this issue was since the print was
never properly completing, you would not see the "Please Remove" pop-up
when the printed ticket was waiting in the chute until the print
timed-out...so you would see the "Ticket Printing" dialog remain
onscreen until the print timed-out instead of being replaced with the
"Please Remove" pop-up when the paper in chute was detected.