TXM-8934 Updating to recover the GEW Bonus Transaction (#659)
Updating to attempt to recover the bonus pay if it is in progress
otherwise let the game recovery pay it normally once it finishes. The
bonus was not recovered because the current pay type was not persisted.
Just call recover internal. If nothing gets recovered, then it will
return a null object and no task, and the game play state will handle
the recovery from there