TXM-9196 - Display "Pending Acceptance" rather than Accepted (#725)
While in the Acceptance state, the Note Acceptor can still fail and
cause the Note/Voucher to be returned.
- At the moment, the Logs will show, and Acceptance state with the
details describing that the note has been Accepted. This is untrue as it
can still be returned. Even though it is just a transient state this
change will ensure that "Pending Acceptance" will be shown instead,
alluding to the fact that the Note Acceptance can still fail.