TXM-9064 Upgrading the GRPC version for address leaks and crashes (#692) (#773)
The bingo project was downgraded with #25 by accident. This was
originally done with 6cff559b and #32. This is re-updating to the
version that fixes the crash as well as upgrading for addressing a
resource leak here: https://github.com/grpc/grpc/issues/8451 and
https://github.com/grpc/grpc/issues/28153.

GRPC core for C# is currently in maintenance mode and only taking
critical fixes.