TXM-8942 Initial Implementation of 3rd Party Keyboard (#790)
First pass at implementing 3rd party keyboard. This is a work in
progress and is not ready to go into the branch. This is being put up
now to get some eyes on the basic implementation. The UI a little work
and we need to get approval for a license for the 3rd party library.

Co-authored-by: Chris Tisdale <christopher.tisdale@aristocrat.com>