TXM-9610 Game configuration should be disabled if game doesn't support enabled languages (#926)
Filter out games that do not support enabled languages everytime enabled
languages are changed.

Ideally a solution is supposed to be making game configuration page
listening to language changed event, refilter games whenever this event
is fired. But it can't be done, because the game configuration is unable
to receive language changed event. Being a subclass of
OperatorMenuPageViewModelBase, the game configuration page by default
unsubscribes all its subscriptions when unloaded. As a result, the
language changed event would never be received by the game configuration
page.

The current way I do is to refilter games based on enabled languages
when the game configuration page is reloaded.

Co-authored-by: Wei Zhou <wei.zhou@aristocrat.com>