TXM-9977 exception 85 reports twice after exit out of audit menu while performing combination test (#1019)
Issue is actually not SAS related but caused by GameDiagnostics making a
duplicate call to the OperatorMenuLauncher's exiting methods. It's a
very close race, but IsShowing is sometimes not set to false by the time
the condition is checked in the GameDiagnostics.Reset method. I suspect
it is due to the WaitHandle (_exitAllowed), but this fix should solve
the issue