TXM-7699 Fixes wrong game names in lobby.config.xml for 99 jurisdictions (#1155)
https://jerry.aristocrat.com/browse/TXM-7699
All changes relating to this jira were fixed in #505. There was an issue
regarding the names that was mentioned during testing and later was
asked of us to fix.

Studio provided names in an email, and I have changed the names of any
games they have mentioned in all jurisdictions that were initially
changed by this Jira.

All changes in this PR are game name changes involving xtraCard games in
the lobby.config.xml in the childeren of
DefaultGameOrderLightningLinkEnabled and
DefaultGameOrderLightningLinkDisabled.