TXM-10888 Refactor Harkey reels control into capabilities (#1262)
Split reel controller code for Harkey reels into different interfaces
based on capabilities provided by the Harkey protocol. This is pre-work
to be able to easily add other reel controllers that offer shared or
different capabilities.

The bulk of the files that were changed only contain namespace changes.
Capability specific code (lighting, brightness, spin) was moved out of
the ReelControllerAdapter and moved into capability specific classes.
The ReelControllerAdapter returns capability implementations to be used
anywhere capability code that was handled by the adapter was called
before. Anywhere that needs a capability will first check to see if the
controller supports it, then it will request an implementation of that
capability. The capability implementations wrap the calls and pass them
along to the particular controller implementation, in this case, the GDS
(Harkey) implementation.