TXM-10998 Remove unused TransferStatusEvent (#1300)
Removed existence of TransferStatusEvent in monaco including tests.

Note: There is a resource string Resources.TransferStatusErrorText that
I would have gotten rid of, but it's still being using by the
LoadingRegionsAndTemplatesEvent and that is getting published in 1
location. Not sure how important this event is either. One event at a
time :)

---------

Co-authored-by: Aaron Oxford <64188123+aaron-oxford@users.noreply.github.com>