TXM-11097 Fixes attract config last row not showing (#1344)
## Issue
There was a hidden header row in the DataGrid in
`AttractCustimizationView.xaml` that was causing the last row to be
offscreen and never being able to be shown.

Although for the data grid we have set `HeadersVisibility="None"`, the
styling for this element was being overridden. In the Xaml file
`Controls.DataGrid.xaml` the `DataGridHeadersVisibilityConverter` was
used to convert the HeadersVisibility into a visible type. In
`DataGridHeadersVisibilityConverter.cs`
[Visibility.](https://learn.microsoft.com/en-us/dotnet/api/system.windows.visibility?view=windowsdesktop-7.0)Hidden
was initially used which takes into account the space of the item but
the fully hidden one is Visibility.Collapsed.

## Solution
There are two changes here,

### DataGridHeadersVisibilityConverter.cs
From Visibility.Hidden changed to Visibility.Collapsed so that there is
no ghost header taking space.

### AttractCustomizationView.xaml
Since the header has now been removed the spacing needs to be adjusted
to not look too close so a top margin of 10 was added.