TXM-10814 - Handling display of multiple cards (#1372)
This is the first pass review to get my changes so far into the R2-SB
branch. My approach was to add a GameIndex property to several events
and split the card-specific data in the UI ViewModel into a dictionary
keyed to GameIndex. There are some things that are still waiting on
other work, but I wanted to get these changes merged so we can all
collaborate on what's left.