TXM-11427 Do not add credits in show mode if hand counts are tracked (#1408)
Problem: If platform is started in show mode, some credits are added to
the machine. If no games are enabled on the machine and hand counts are
enabled for the jurisdiction, there is no way to open the audit menu and
enable the games as cashout won't happen when hand count is zero.
Solution: In show mode, check if jurisdiction does not have hand counts
enabled or the EGM has any enabled game before adding the credits.