TXM-10825 Merge TXM-10807 from 2.7 (#1398)
TXM-10807 Single game button image was not being scaled (#1233)

https://jerry.aristocrat.com/browse/TXM-10825

In addition to merging this from 2.7 branch, I removed a bunch of the
converters that were just doing the work of Viewboxes or adding extra
margin calculations that were not necessary. Now the landscape lobby is
one large Viewbox that will resize to the available screen without
resizing every individual item. This should make maintenance and layout
changes easier going forward. I would have done this on 2.7 as well but
we were already close to code complete and I didn't want to make huge
changes. However, now that we are redoing the lobby anyway per HRG, a
lot of this will be going into 2.7 anyway. So it goes.