TXM-9696 Race Condition with Note Acceptor error and SAS (#1297)
New pull request replacing #1187 that was merging into release/2.7
instead of main.

This contains the new set of changes to fix the race condition where the
Recover() method of the CurrencyInProvider never ran, causing the
pending log to never be updated after a power failure and be left as
default (Pending and Accepted). This fix moves some of the logic from
the OnRun() method to the OnInitialize() method to fix this race
condition.