TXM-10815 - Disable card instances on the Bingo overlay for games that are no longer in play (#1523)
Definitions:
- Game configuration: The XML overlay configuration (including card
instance settings) and CSS scenes set by the game
- Card/Card instance/instance: the graphical representation of a single
bingo card on the overlay, not to be confused with "card" in the
gameplay sense
- Card states: The possible states of each card instance specified in
the game configuration; see below

Card States:
- Enabled (!Disabled): Card instances have the `Enabled` state when they
are in play. Enabled instances are included in the card cycle and
displayed according to the game configuration (refer to the `Inactive`
state below). Disabled instances are not in play; their visibility and
appearance is determined by the game configuration—specifically the
`disabled-card` CSS class and the `HideDisabled` card instance setting
in the XML. Disabled card instances will automatically have the `Hidden`
state applied if the `HideDisabled` setting is true for that instance.
If the class is not defined and `HideDisabled` is false, the card
appearance will not change.
- Hidden: A hidden card instance has the `hidden-card` CSS class
applied. The exact behavior is determined by this class and any defined
animations, but the intended result is to render a card invisible to the
player.
- Inactive: Card cycling makes one enabled card instance active at a
time based on the interval set in the game configuration. When an active
card instance has patterns, the pattern cycle runs while the instance is
active, and the instance remains active until the pattern cycle
completes. Inactive card instances will automatically have the `Hidden`
state applied if the `HideInactive` card setting is true.

Current functionality: 
- Any configured cards beyond the main bingo card start as Disabled on
launch.
- The `BingoGameNewCardEvent` is fired for each card in the outcomes;
the game index for the outcome is passed in the event. This event can
also be fired during recovery.
- The `BingoGameNewCardEvent` handler sets the Enabled state for the
card instance associated with the game index and sets the numbers on the
card. This is not currently built to handle more than one card per game
index.
- The `GamePlayInitiatedEvent` handler clears the numbers from all bingo
cards. The card states are not changed by this handler.
- There are currently no events that will apply the `Disabled` state to
a card once it has been enabled.

**The above documentation is available at
https://confy.aristocrat.com/display/MON/Bingo+Overlay**

Proposed change:
This change sends an event to the overlay to apply the `Disabled` state
to any card instance that is not updated by an outcome; i.e. if the side
bet game is opted out when a game is played. The state of the side bet
window itself is not considered. The goal is to ensure this state is set
when a card is not in play; how it affects the card will be determined
by the studio (see above).

---------

Co-authored-by: stephan.lemon <stephan.lemon@aristocrat.com>
Co-authored-by: Steve Meyer <Steve.Meyer@Aristocrat.com>
Co-authored-by: torencarr <toren.carr@aristocrat.com>