TXM-11662 Added support to close dynamic help on any door open (#1532)
This was found by BMM for LS and deemed a minor issue, so only
addressing it for next release from trunk. This simply exits dynamic
help for any door is open event so that the resulting door open lock-up
is not hidden behind the dynamic help view which could result in the
unlikely event of a door is open while tapping on/navigating thru the
dynamic help display.