TXM-11813 - Keep mechanical buttons lit when Cashout If residual Credits are enough to play game (#1591)
**Issue**

When Cashout happens in COAM the platform overrides the lamps on the
mechanical button deck and turns off all lamps.
This is because we assume that cashout removes all credits from the
machine.
In COAM, we may cashout and still have enough credits to play a game.

**Solution**
Before turning off all lamps, check if we have enough residual credits
to play a game. If we do, then do not turn off all the lamps. If we
don't then continue in normal behavior.
This **change** will only affect COAM and will only affect the buttons
when we are in the Game state.