TXM-11874 - Configuration Overrides crashing platform (#1606)
**Issue**
Locale.Config.xml was not found using the **Specific jurisdiction key**.
There was also changes to Locale.config.xml which were not in our COAM
branch.

**Solution**
We were not persisting the Jurisdiction key correctly (Fixed). Changed
the Locale.config.xml to be the same as the others in the platform. As
well as to adhere to the new Override behavior.