TXM-11824 Update cross game progressive functionality to function with the 1.1.10 protocols (#1625)
The fake progressive server currently only supports registering a single
progressive game so the platform code is only sending a single sub game
until this is fixed.
This fixes both registration and progressive updates. Claim and Award
code is not tested at this time.
After an nvram clear the progressive system was failing to register
since there was no active game. Had to do a lot of refactoring to get
this fixed.