TXM-11931 - Removed Fractional Amount from the progressive level (#1711)
**Problem**

Progressive levels shown in the Lobby Icon for that game were rounding
the value up to the closest cent: (See images Below)
The real progressive amount was shown differently when in the game.

![image](https://github.com/Aristocrat-Monaco-Platform/monaco/assets/38898894/d34924f3-eb9f-442c-aee2-7243457db09c)

![image](https://github.com/Aristocrat-Monaco-Platform/monaco/assets/38898894/72af2095-8225-4da2-a3f4-7a95ee77f4d7)

**Solution**
When a progressive level is paid, the fractional amount is taken away
and added to the overflow amount. Calculate the value by taking away the
fractional amount will display the correct value.