TXM-11589 - Do not return to Lobby from Lockup in game/gameLoading (#1773)
**Problem One**
Credit Limit Banner and subsequent Printing Dialog lockups are causing
the Lobby to transition from Disabled state to GameLoaded State when
platform is already in the game, and the game is running. The Transition
from Disabled to GameLoaded is a transition which causes the Platform to
return to Lobby. We do not want to return to Lobby, especially due to
lockups.

**Problem Two**
After clicking a game from the Lobby, while the game is Loading, if a
lockup occurs frequently then the platform will return to lobby.

**Solution**
Only send GameLoaded Trigger if we are not in the Disabled State. This
way we do not return to lobby.