TXM-10874 Persist IContinuationContext after commits for wager matches (#1895)
When we are issuing wager matches which are less than the wager match
total award, we know to stop awarding when we transition gameplay state
to idle after the initial award as part of WagerMatch.CanPay (the bonus
state will still be pending as we haven't reached the wager match
total). However, sometimes this gameplay state transition happens too
slowly and we attempt to pay a wager match for a second time. Usually
this will fail when we hit CanPay for a second time as referenced by the
Pay method so it ends up not mattering, but not always. When this
happens, we can attempt to pay a wager match for a second time at the
end of the game round.

To fix this, I made sure that IContinuationContexts can be persisted
across commits. That way, when the next commit happens and encounters
this condition, it will know that the wager match has already been fully
paid and stop anyway.