TXARCH-1211 Cannot setup Progressives in the latest RTP builds (#1982)
This change fixes the inability to set up progressive levels, the exact
change that caused this defect is unknown however, it was a result of
the ProgressiveLevelProvider being populated with the first set of
ProgressiveLevels after the ProgressiveConfigurationProvider attempted
to grab them to set up the first games configurable level. To fix this
issue the publishing of the GameAddedEvent was moved to be called after
the ProgressiveLevelProvider was able to load the progressive levels in
order to allow them to be present when it was necessary for them to be.