TXARCH-1247 - batch value updates to LinkedProgressiveLevels
We experimented with filtering out level updates when there was no
change in the Amount field (i.e. the level hasn't changed), but opted to
allow those back in just to capture the updated ProgValueSeq field. We
can revisit this in the future if needed. One option would be to filter
out level updates until the persisted ProgValueSeq minus the new
ProgValueSeq exceeded a threshold.

So the real change here is to cache all level updates for 5 (arbitrary)
seconds and then update all of the levels at once, rather than via
separate updates every 2 seconds for every level.