TXM-12785 - Adds a delay for payout to ensure CashOutButton shows before payout in Audit menu logs. (#2130)
https://jerry.aristocrat.com/browse/TXM-12785

A simple `Task.Delay()` for one second was added to ensure payout was
persisted after `CashOutButtonEvent` was persisted.

`CashOutButtonEvent` is fired before payout but the persistence for
general events works on a timer, where on timer elapsed it would gather
all the events that occurred and persisted them. This would cause the
`CashOutButtonEvent` to appear after payout occurring in the audit menu
logs.