TXM-12708 Cannot export machine settings (#2191)
Issue: Cannot export machine settings. The rootcause was in the
CultureProvider.cs we are throwing the exception if resource is null so
we never get the value of resource = null in the getter of
HardMeterMapSelectionValue in MachineSettings.cs

Solution: If the resource is null do not throw the exception instead
return the null value. If the value is null then it takes
_hardMeterMapSelectionValue
get => OperatorLocalizer.GetString(_hardMeterMapSelectionValue) ??
_hardMeterMapSelectionValue;