TXM-13256 Sound is getting played on clicking Max bet button even when it is unlit and non-functional on lobby (#2321)
Issue: Georgia COAM Sound is getting played on clicking Max bet button
even when it is unlit and non-functional on lobby

Solution: When the max bet button is clicked there must be no sound
played as its unlit and non functional in the lobby. Added a check which
decided whether to play sound or not.