TXM-13097 The top screen do not display the game top screen when cashout (#2295)
**Note: This PR is marked Ready for Review

Issue: The top screen do not display the game top screen when cashout.
It was created when Quebec team removed the default behavior of how we
show top screen with game during cashout.

Solution: In all other jurisdiction Top screen will display the game
screen when we cashout. For Quebec jurisdiction top screen should not
display the game. So added the the check like before while we set value
for IsTopScreenRenderingDisabled.