TXM-13327 The top screen do not display the game top screen when … (#2360)
**Note: Issue merge into main**

Issue: The top screen do not display the game top screen when cashout.
Bug was created when Quebec team removed the default behavior of how we
show top screen with game during cashout.

Solution: In all other jurisdiction Top screen will display the game
screen when we cashout. For Quebec jurisdiction top screen should not
display the game. So added the the check like before while we set value
for IsTopScreenRenderingDisabled.