TXM-13317 Sound is getting played on clicking Max bet button even when it is unlit and non-functional on lobby (#2357)
**Note: COAM Issue merge into main**

Issue: Georgia COAM Sound is getting played on clicking Max bet button
even when it is unlit and non-functional on lobby

Solution: When the max bet button is clicked there must be no sound
played as its unlit and non functional in the lobby. Added a check which
decides whether to play sound or not.