TXM-10993 Added local pacakge and nuget configuration details (#2399)
We have created new local Cabinet package which will provide the Neptune
Single Support in OS, to provide Neptune Single cabinet Support in
Monaco Platform we have done these changes.

1. Added Nuget config to support local package upgrade for cabinet
package.
2. Updated the Package reference to support Neptune cabinet.

**NOTE:** We are using this local package because we don't have official
package as of now.
These are the local changes for this branch only, these changes will not
be merged into the main. When we will be merging it in the main, Using
TXM-13560, we will remove nuget.config file and upgrade to the official
package reference.