TXM-11765 Support Lesotho Loti currency  (#2386)
Lesotho currency in Windows is South African Rand rather than Loti,
which is required in Monaco. To support this, Loti currency format is
added to CurrencyDefaults.xml. Monaco uses the configuration to create
the Currency object and use the configured format to display currency in
the Lobby, Operator menu and Games.

This also includes
- the fix if no BNA is configured
- patch merged from TXM-12240
- update unit tests