TXM-13504 - Create meters for LinkedProgressiveLevels (#2350)
Create WageredAmount and PlayedCount meters tracked for
LinkedProgressiveLevels. These meters are incremented anytime an
associated ProgressiveLevel's equivalent meters are. These meters are
sent to the G2S Progressive host (vertex) instead of a specific game's
meters to avoid the risk of meter rollback when multiple games are
contributing to the same progressive pool