TXM-13505 - Prevent G2S permanently disabling games when it's not the validation protocol (#2376)
This changes G2S to enable any devices assigned to the EGM by default.
When G2S is not the validation protocol, GamePlay devices will instead
be assigned to the EGM, and the EGM will now enable those devices. This
prevents the devices from being disabled until a validation host (that
is never coming) enables them.