TXM-13232 Use RuntimeFlagHandler to update the cashing out flag value (#2516)
From Satya's investigation, platform doesn't update the CashingOut flag
in the second time of cashout.

When player tries to cashout in the game, GDK will send cashout command
to platform. Then Cashout command handler will call RuntimeFlagHandler
to update the CashingOut flag. The RuntimeFlagHandler has cached current
state and check if the flag value has been changed before calling
runtime proxy to update the flag. In the CashoutDialogViewModel, it
calls runtime proxy directly, which will not update the cached value in
the RuntimeFlagHandler. When player tries the cashout in the second
time, RuntimeFlagHandler doesn't call runtime proxy as it finds the flag
value is not changed.

The fix is to use RuntimeFlagHandler in the CashoutDialogViewModel.