Making the AttractState.Videos Immutable again
This change primarily is to make the AttractState.Videos field Immutable again, it was changed as at the time I didnt have a work around in mind, I do now and it is best to keep it as an immutable collection.

Also reverted the name change to the AttractSetVideosAction, after further inspection of old lobby code this action should set the collection as opposed to adding to it.