TXM-13140 Only allow for speaker devices to cause sound tilt (#2556)
Fleshed out the audio device detection so that we can grab the device
descriptor from each audio device. From that, forced detection to only
apply to speakers so that when we restart an EGM, it doesn't simply
default to the Nvidia HDMI audio (or whatever other non-speaker audio
device is connected). This will allow the audio disconnected lock up to
actually work properly on startup on an EGM.