TXM-13881 TXM-13829 Synchronize IPC start/stop with game process start (#2485) (#2638)
With 455e7c6c88d2c852af792c56a317d2826a2f0b8e, we added the ability to
shut down the SNAPP server when the game is exited. However, if we open
another game extremely quickly after entering the lobby, it is possible
to get into a state where we launch the SNAPP server for the purposes of
communicating with the new game before we get to the point where we shut
down the SNAPP server after the first game's process exits. This change
synchronizes IPC start/stop with the game process start so that when we
attempt to shut down the SNAPP server, we know we at least aren't about
to launch a new game after starting the SNAPP server. Then, we can check
to see if we have launched a new game to determine if we really should
shut down the SNAPP server or not.

(cherry picked from commit 85224560fc82bc48090cc82f929f882838fd993c)