TXARCH-1315: Doors would not trigger opened/closed on Mars Upright & Helix. (#2644)
- Issue stems from discrepancy that occurs when door open/close event is
timestamped before initial configuration of the QC Inspection tool where
operator sets the current date and time. If the Windows installation for
QC Inspection Tool has a date/time set in the future when the above
occurs, the DoorService does not correctly process open/close door
events because it compares the timestamp of the current door action with
the last recorded door action to decide whether to publish the
open/close event.

- Added a conditional check for Inspection mode in DoorService to
address this issue. Update will ensure the system correctly logs and
updates the last open/close event timestamp when in Inspection mode.