TXARCH-1485 - don't restart commsDisabled timer when in online state
Add a clause to the CommsDisabledAck handling that avoids restarting the commsDisabled timer if the CommunicationsDevice has already transitioned to online state.
This avoids a scenario where Monaco thinks it is online, but is internally trying to start the sync process again.
This can lead to a communications deadlock that will never recover.