More frontend updates
Added method to the IAttractService so casting would not be needed.

Added an effect handler for AttractVideoNextAction to allow for the video to progress to the next attract index upon completion.

Added some extra selectors to allow the front-end to update properly when state values are updated.

Did some better formatting in the AttractMainView to improve readability.

(DesignHeight value on the AttractMainView.xaml will be updated to change within the xaml.cs file similar to how it functions in the current lobby)