Some more front-end updates
Fixed an issue resulting in the attract videos always using the default attract videos even when the game(s) have specific videos that should be played. This was caused by the SetAttractVideos action never being called.

Removed the IsBottomAttractVisible flag from the view model and updated the view to utilize the IsBottomAttractVideoPlaying flag instead, since that will be false if the video is not visible. This may be reverted in the future if I am misunderstanding how attract works.

Also fixed where I set the Height/Width on the attract view, I accidentally swapped the design height and width rather than the Bink Controllers height and width, visual quality looks much better now. That will obviously be made dynamic in the future.